feat: Gemini 3 preview support
This was harder than it needed to be. Gemini 3 preview requires to have
a "thought_signature" to work. This is added via the field
"extra_content" which apparently is openai API compatible. Fairly straightforward
to add.

But..! As it is an experimental model, there seem sto be a bug on their
end where once the tool-calls have been completed, gemini-3-preview
sends an additional erroneous tools-call for no reason whatsoever. The
"proof" that it's on their end can be seen via checking the
tools-call-id, and noting that it's not been sent before: it's both set
and not been sent before = not a caching issue.

So to handle this case, I needed to add additional checks in the generic
stream querier which detects if the underlying model most likely is
gemini-3-preview (one of the tool-calls has had a thought_signature) and
then return early if a subsequent call is without thought_signature.
Sigh.

Now I can finally get back to doing what I was supposed to do: hack on a
dart project...

Additional features:
        * Tightneted up security by:
                * Removed mcp everything printEnv (this kept sending all
                  my access tokens to the models I was testing out)
                * Removed any api-key printing from DEBUG=1
        * Tests written by gemini-3-pro-preview to be a bit meta